pyproject.toml
[project]
name = "lot-mcu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "websocket-client>=1.9.0",
]
